passes reg tests

+++ backend/db/schedules.py
@@ -4,6 +4,7 @@ from db.couchbase_server import *
 import couchbase.subdocument as subdoc 
 from adb_utils import catch_missing, require_json_data, catch_already_exists, json_response, SubdocPathNotFoundError
 from db.offerings import offeringGet
+import db.registration as registration
 
 
 sched_bucket = cluster.open_bucket('schedules')
@@ -28,13 +29,13 @@ def pull_enrollments(schedules):
 
 
 def del_all_scheds_for(username):
-    from db.registration import unregister  # TODO: guessing there would be an import circle otherwise. Should refactor to get rid of that.
+    # from db.registration import unregister  # TODO: guessing there would be an import circle otherwise. Should refactor to get rid of that.
     schedules = all_schedules_for(username)
     # de-register from all classes
     enrollments = pull_enrollments(schedules)
     for enrollment in enrollments:
         try:
-            unregister(username, *enrollment)
+            registration.unregister(username, *enrollment)
         except:
             pass
     # delete all schedules
@@ -44,6 +45,8 @@ def del_all_scheds_for(username):
 
 
 def __unsafe_delete_user_schedules(username, *quarters):
+    if not quarters or not len(quarters):
+        return True
     sched_keys = list(username+'-'+quarter for quarter in quarters)
     sched_bucket.remove_multi(sched_keys, quiet=True)
     return True
